Rename locations panel index file to locations container
