+++ src/shared/LocationsPanel/LocationsContainer.js
